vector lib rtree segfault bugfix
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@36003 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ lib/vector/rtree/rect.c
@@ -354,7 +354,7 @@ int RTreeContained(struct Rect *R, struct Rect *S)
     register struct Rect *r = R, *s = S;
     register int i, j, result;
 
-    assert((int)r && (int)s);
+    assert(r && s); /* same as in RTreeOverlap() */
 
     /* undefined rect is contained in any other */
     if (Undefined(r))
